FIX/lib/StupidArtnet.py: ensure BUFFER length equals PACKET_SIZE
When constructing a StupidArtnet instance the `self.BUFFER` is setup using the
argument `packet_size` instead of the `self.PACKET_SIZE` property, which could
lead to mismatch of the buffers length and the requested packet size. This happened
because `self.PACKET_SIZE` is initialized using the `put_in_range` function,
which is using the `make_even` argument set to True by default. Passing an
uneven packet size to the constructor therefore leads to mistches, e.g.:

```python
led_count = 9
artnet = StupidArtnet("127.0.0.1", 0, led_count * 3, 24)
```

The fix was easy by initializing `BUFFER` using the `PACKET_SIZE` property
instead of the passed arugment :)